requirements.txt
fastapi
uvicorn[standard]
pydantic
passlib[bcrypt]
python-jose[cryptography]
pandas
numpy
matplotlib
scikit-learn